Refactor compile_profiles.py controls logic
Allow other scripts to use controls-handling logic, exporting it from
build-scripts/compile_profiles.py into ssg/. This will allow refchecker
to reuse this to check profiles with control files for references.
